game/app.js > main.js
