--- backend/game/app.js
@@ -1,71 +1,0 @@
-const express = require("express");
-const app = express();
-
-// --- configs
-
-const PORT = process.env.PORT || 3000;
-app.use(express.json())
-
-// --- imports
-
-const xo = require("./site")
-const games = new xo();
-
-// --- frontend
-
-app.use(express.static('../frontend'));
-
-// --- Endpoints
-
-app.post('/game', (req, res) => {
-    
-    console.log('/game');
-    console.log("req", req.body);
-    let rs = games.request_game(req.body);
-    console.log("res", rs);
-    console.log();
-    res.send(rs);
-    
-});
-
-app.post('/play', (req, res) => {
-    
-    console.log('/play');
-    console.log("req", req.body);
-    let rs = games.request_play(req.body);
-    console.log("res", rs);
-    console.log();
-    res.send(rs);
-    
-});
-
-app.post('/sync', (req, res) => {
-    
-    console.log('/sync');
-    console.log("req", req.body);
-    let rs = games.request_sync(req.body);
-    console.log("res", rs);
-    console.log();
-    res.send(rs);
-    
-});
-
-app.post('/move', (req, res) => {
-    
-    console.log('/play');
-    console.log("req", req.body);
-    let rs = games.request_move(req.body);
-    console.log("res", rs);
-    console.log();
-    res.send(rs);
-    
-});
-
-
-
-// --- Runs
-
-app.listen(PORT, async (error) => {
-    if (error) { console.log(error); }
-    console.log(`Server listening on port ${PORT}`);
-});
